requirements.txt
python-dotenv
langchain
langchain-openai
requests
beautifulsoup4
azure-identity
azure-search-documents
langchain-community
chardet
